mod4/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
numpy
tqdm
realesrgan        # if pip package is available; otherwise adapt to local wrapper
fhirclient        # SMART on FHIR client, alternative use requests
requests
python-jose[cryptography]  # for JWT handling if needed
python-multipart
